backend/requirements.txt
Flask
flask-cors
requests
gspread
google-auth
python-dotenv